+++ src/grip/index.tsx
@@ -1,0 +1,1 @@
+export { GripProps, Grip } from './Grip'