+++ src/text/story.tsx
@@ -15,7 +15,7 @@ stories.add('Default', () => (
       tag={select('Tag', TextTagType, TextTagType.SPAN)}
       textColor={select('Color', color, 'primary')}
       newlineToBr={boolean('newlineToBr', true)}
-      ariaLabel={text('Aria label', 'Aria label')}
+      aria-label={text('Aria label', 'Aria label')}
     >
       {text('Text', 'This is an example')}
     </Text>